llamaindex/requirements.txt
llama-index
llama-index-llms-openai
python-dotenv 
llama_deploy